requirements.txt
# Core Flask
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0

# Database
mysqlclient==2.2.1
SQLAlchemy==2.0.23

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# AI / ML
opencv-python>=4.8.1
numpy==1.26.2
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2

# NLP
nltk==3.8.1
spacy==3.7.2

# Utilities
Pillow==10.1.0
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Ollama (local LLM)
requests>=2.31.0
